classwork2
In MongoDB, aggregation operations process the data records/documents and return computed results. It collects values from various documents groups them together and then performs different types of operations on that grouped data like sum, average, minimum, maximum, etc to return a computed result.
Aggregation pipeline:
In MongoDB, the aggregation pipeline consists of stages and each stage transforms the document. Or in other words, the aggregation pipeline is a multi-stage pipeline, so in each state, the documents taken as input and produce the resultant set of documents now in the next stage(id available) the resultant documents taken as input and produce output, this process is going on till the last stage.
$match: It is used for filtering the documents and can reduce the amount of documents that are given as input to the next stage.
$project: It is used to select some specific fields from a collection.
$group: It is used to group documents based on some value.
$sort: It is used to sort the document that is rearranging them
$out: It is used to write resulting documents to a new collection
$limit: It is used to pass the first n number of documents thus limiting them.
$set: It replaces the value of a field with the specified value.
$count: It returns a count of the remaining documents in the aggregation pipeline and assigns the value to a field called passing_scores (or)to count the documents.